:recycle: myPagePage refactoring
